split persistent ui into multiple files
