+++ src/models/stores/persistent-ui/persistent-ui.ts
@@ -1,7 +1,6 @@
 import {
   getSnapshot, applySnapshot, types, onSnapshot,
-  SnapshotIn, Instance, cast
-} from "mobx-state-tree";
+  SnapshotIn, Instance} from "mobx-state-tree";
 import { AppConfigModelType } from "../app-config-model";
 import {
   DocFilterType, DocFilterTypeEnum, kDividerHalf, kDividerMax,
@@ -23,211 +22,10 @@ import { urlParams } from "../../../utilities/url-params";
 import { removeLoadingMessage, showLoadingMessage } from "../../../utilities/loading-utils";
 import { SortedDocuments } from "../sorted-documents";
 import { cloneDeep } from "lodash";
+import { UITabModel, UITabModel_V1 } from "./ui-tab-model";
 
 export const kPersistentUiStateVersion2 = "2.0.0";
 export const kPersistentUiStateVersion1 = "1.0.0";
-
-export const UIDocumentGroup = types
-  .model("UIDocumentGroup", {
-    id: types.identifier,
-    /**
-     * Either undefined, an empty array, or an array of document ids.
-     * - Undefined means the user hasn't chosen anything yet so we can show a default view
-     *   in some cases this might be showing the first document of the listing.
-     * - Empty array means the user has explicitly chosen to close the documents and
-     *   this typically means they'll be viewing the listing of documents.
-     * - The array typically has a single document id, in some cases a second document
-     *   can be open at the same time.
-     *
-     * TODO: another approach for handling the "user hasn't chosen anything yet" case is
-     * to look to see if there is a document group at all. Currently the visitedDocumentGroup
-     * isn't created until a user actually opens a document. But before making that switch
-     * we need to have a working example of this "automaticallyOpenFirstDocument" setting.
-     */
-    currentDocumentKeys: types.maybe(types.array(types.string))
-  })
-  .views(self => ({
-    get primaryDocumentKey() {
-      if (!self.currentDocumentKeys || self.currentDocumentKeys.length < 1) {
-        return undefined;
-      }
-      return self.currentDocumentKeys[0];
-    },
-    get secondaryDocumentKey() {
-      if (!self.currentDocumentKeys || self.currentDocumentKeys.length < 2) {
-        return undefined;
-      }
-      return self.currentDocumentKeys[1];
-    },
-    get userExplicitlyClosedDocument() {
-      return self.currentDocumentKeys?.length === 0;
-    }
-  }))
-  .actions(self => ({
-    setPrimaryDocumentKey(documentKey: string) {
-      let docKeys = self.currentDocumentKeys;
-      if (!docKeys) {
-        self.currentDocumentKeys = cast([]);
-        docKeys = self.currentDocumentKeys!;
-      }
-
-      if (docKeys.length === 0) {
-        docKeys.push(documentKey);
-      } else if (docKeys.length > 0) {
-        docKeys[0] = documentKey;
-      }
-    },
-    /**
-     * If there is no primary document we just set the documentKey as the primary document
-     * and print a warning in the console
-     * @param documentKey
-     */
-    setSecondaryDocumentKey(documentKey: string) {
-      let docKeys = self.currentDocumentKeys;
-      if (!docKeys) {
-        self.currentDocumentKeys = cast([]);
-        docKeys = self.currentDocumentKeys!;
-      }
-
-      if (docKeys.length === 0) {
-        console.warn("setting a secondary document when there is no primary document");
-        docKeys.push(documentKey);
-      } else if (docKeys.length === 1) {
-        docKeys.push(documentKey);
-      } else if (docKeys.length > 1) {
-        docKeys[1] = documentKey;
-      }
-    },
-    /**
-     * Close the current primary document. If there is a secondary document it will become
-     * the primary document.
-     * @returns
-     */
-    closePrimaryDocument() {
-      const docKeys = self.currentDocumentKeys;
-      if (!docKeys) {
-        // The user is taking the explicit action to close the document
-        // So we save that as an empty array.
-        self.currentDocumentKeys = cast([]);
-        return;
-      }
-
-      if (docKeys.length === 0) return;
-
-      docKeys.splice(0, 1);
-    },
-    /**
-     *
-     * @returns
-     */
-    closeSecondaryDocument() {
-      const docKeys = self.currentDocumentKeys;
-      if (!docKeys || docKeys.length < 2) return;
-
-      docKeys.splice(1, 1);
-    }
-  }));
-
-/**
- * This model is used to track which document group of a tab is open, which documents are open in
- * each document group, and which secondaryDocuments (comparison) documents are open in each
- * document group.
- * For the StudentGroupView the currentDocumentGroupId is the student group id.
- * For the SortWorkView the currentDocumentGroupId in an encoding the filters for group documents
- * that are currently open.
- *
- * TODO: Create different tab types so so the models can more clearly match the UI
- */
-export const UITabModel = types
-  .model("UITab", {
-    id: types.identifier,
-    /**
-     * The currently open document group on this tab. This could be a subtab, a workgroup,
-     * or a collapsible section on the sort work tab.
-     */
-    currentDocumentGroupId: types.maybe(types.string),
-    /**
-     * Document groups that the user has visited.
-     */
-    visitedDocumentGroups: types.map(UIDocumentGroup),
-  })
-  .views(self => ({
-    /**
-     * Return the document group. Note: if there is no state saved about this document
-     * group then undefined will be returned. Use getOrCreateDocumentGroup if you need to save
-     * state.
-     * @param docGroupId
-     * @returns
-     */
-    getDocumentGroup(docGroupId: string) {
-      return self.visitedDocumentGroups.get(docGroupId);
-    },
-  }))
-  .views(self => ({
-    /**
-     * Note: If there hasn't been any state saved about the currentDocumentGroupId then
-     * this will return undefined. Use getOrCreateDocumentGroup if you need to save
-     * state.
-     */
-    get currentDocumentGroup() {
-      if (!self.currentDocumentGroupId) return undefined;
-
-      return self.getDocumentGroup(self.currentDocumentGroupId);
-    },
-    getPrimaryDocumentInDocumentGroup(docGroupId: string) {
-      return self.visitedDocumentGroups.get(docGroupId)?.primaryDocumentKey;
-    },
-    getSecondaryDocumentInDocumentGroup(docGroupId: string) {
-      return self.visitedDocumentGroups.get(docGroupId)?.secondaryDocumentKey;
-    }
-  }))
-  .actions(self => ({
-    /**
-     * This should not be used from view. Because this is an action the properties it reads
-     * will not be observed by that parent view. For example if the documentGroup is
-     * deleted, that would not trigger the parent view to be re-rendered.
-     * @param docGroupId
-     * @returns
-     */
-    getOrCreateDocumentGroup(docGroupId: string) {
-      let group = self.visitedDocumentGroups.get(docGroupId);
-      if (!group) {
-        group = UIDocumentGroup.create({id: docGroupId});
-        self.visitedDocumentGroups.put(group);
-      }
-      return group;
-    }
-  }))
-  .actions(self => ({
-    /**
-     * This will create or update the UIDocumentGroup in this tab.
-     * This will not change the currentDocumentGroup.
-     *
-     * @param docGroupId
-     * @param documentKey
-     */
-    setPrimaryDocumentInDocumentGroup(docGroupId: string, documentKey: string) {
-      const group = self.getOrCreateDocumentGroup(docGroupId);
-      group.setPrimaryDocumentKey(documentKey);
-    },
-    setSecondaryDocumentInDocumentGroup(docGroupId: string, documentKey: string) {
-      const group = self.getOrCreateDocumentGroup(docGroupId);
-      group.setSecondaryDocumentKey(documentKey);
-    }
-  }))
-  .actions(self => ({
-    openPrimaryDocumentInDocumentGroup(docGroupId: string, documentKey: string) {
-      self.setPrimaryDocumentInDocumentGroup(docGroupId, documentKey);
-      self.currentDocumentGroupId = docGroupId;
-    },
-  }));
-
-interface UITabModel_V1 {
-  id: string,
-  openSubTab?: string,
-  openDocuments: Record<string, string>,
-  openSecondaryDocuments: Record<string, string>
-}
 
 export const PersistentUIModelV2 = types
   .model("PersistentUI", {
